fraud_detection/requirements.txt
grpcio==1.70.0
grpcio-tools==1.70.0
protobuf==5.26.1
watchdog==6.0.0
pandas==2.1.3
joblib==1.3.2 
